+++ e2e/cases/manifest/custom-path/src/index.ts
@@ -1,0 +1,3 @@
+import React from 'react';
+
+console.log('hello!', React);
