+++ e2e/cases/manifest/filter/index.test.ts
@@ -1,0 +1,69 @@
+import { expect, getFileContent, rspackTest, test } from '@e2e/helper';
+
+test('should allow to filter files in manifest', async ({ build }) => {
+  const rsbuild = await build({
+    config: {
+      output: {
+        manifest: {
+          filter: (file) => file.name.endsWith('.js'),
+        },
+        filenameHash: false,
+      },
+      performance: {
+        chunkSplit: {
+          strategy: 'all-in-one',
+        },
+      },
+    },
+  });
+
+  const files = rsbuild.getDistFiles();
+
+  const manifestContent = getFileContent(files, 'manifest.json');
+  const manifest = JSON.parse(manifestContent);
+
+  // index.js
+  expect(Object.keys(manifest.allFiles).length).toBe(1);
+
+  expect(manifest.entries.index).toMatchObject({
+    initial: {
+      js: ['/static/js/index.js'],
+    },
+  });
+});
+
+rspackTest(
+  'should allow to include license files in manifest',
+  async ({ build }) => {
+    const rsbuild = await build({
+      config: {
+        output: {
+          manifest: {
+            filter: () => true,
+          },
+          filenameHash: false,
+        },
+        performance: {
+          chunkSplit: {
+            strategy: 'all-in-one',
+          },
+        },
+      },
+    });
+
+    const files = rsbuild.getDistFiles();
+
+    const manifestContent = getFileContent(files, 'manifest.json');
+    const manifest = JSON.parse(manifestContent);
+
+    expect(Object.keys(manifest.allFiles).length).toBe(3);
+
+    expect(manifest.entries.index).toMatchObject({
+      initial: {
+        js: ['/static/js/index.js'],
+      },
+      html: ['/index.html'],
+      assets: ['/static/js/index.js.LICENSE.txt'],
+    });
+  },
+);
