+++ e2e/cases/manifest/filter/src/index.ts
@@ -1,0 +1,3 @@
+import React from 'react';
+
+console.log('hello!', React);
