+++ e2e/cases/manifest/node/rsbuild.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from '@rsbuild/core';
+
+export default defineConfig({
+  output: {
+    target: 'node',
+    manifest: true,
+    filenameHash: false,
+  },
+});
